pyproject.toml
[project]
name = "llm-labs"
version = "0.1.0"
description = "Labs of course NLP and LLM"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "marimo>=0.17.8",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
lab1 = [
    "arxiv>=2.3.1",
    "ddgs>=9.9.1",
    "langchain>=1.0.7",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.3",
    "langgraph>=1.0.3",
]
